chatlings/scripts/requirements.txt
# Python requirements for Chatlings image generation

psycopg2-binary>=2.9.9
perchance>=0.0.1
aiofiles>=23.2.1
aiohttp>=3.13.2
playwright>=1.55.0
python-dotenv>=1.0.0
